third_party/blink/web_tests/external/wpt/fetch/api/request/request-disturbed-expected.txt

This is a testharness.js-based test.
PASS Request's body: initial state
PASS Request without body cannot be disturbed
PASS Check cloning a disturbed request
PASS Check creating a new request from a disturbed request
FAIL Check creating a new request with a new body from a disturbed request Failed to construct 'Request': Cannot construct a Request with a Request object that has already been used.
FAIL Input request used for creating new request became disturbed assert_equals: body should not change expected object "[object ReadableStream]" but got object "[object ReadableStream]"
FAIL Input request used for creating new request became disturbed even if body is not used assert_equals: body should not change expected object "[object ReadableStream]" but got object "[object ReadableStream]"
PASS Check consuming a disturbed request
PASS Request construction failure should not set "bodyUsed"
Harness: the test ran to completion.
